+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/junit/TigerExtension.java
@@ -55,6 +55,7 @@ public class TigerExtension
     return context
         .getTestMethod()
         .map(m -> m.getAnnotation(TigerTest.class))
+        .or(() -> context.getTestClass().map(c -> c.getAnnotation(TigerTest.class)))
         .orElseThrow(
             () ->
                 new TigerEnvironmentStartupException(